+++ src/Firebase/Contract/Auth.php
@@ -27,7 +27,6 @@ use Kreait\Firebase\Exception\Auth\RevokedIdToken;
 use Kreait\Firebase\Exception\Auth\UserDisabled;
 use Kreait\Firebase\Exception\Auth\UserNotFound;
 use Kreait\Firebase\Request;
-use Kreait\Firebase\Value\ClearTextPassword;
 use Lcobucci\JWT\Token;
 use Psr\Http\Message\UriInterface;
 use Traversable;
@@ -84,7 +83,7 @@ interface Auth
 
     /**
      * @param \Stringable|string $email
-     * @param ClearTextPassword|string $password
+     * @param \Stringable|string $password
      *
      * @throws Exception\AuthException
      * @throws Exception\FirebaseException
@@ -116,7 +115,7 @@ interface Auth
 
     /**
      * @param \Stringable|string $uid
-     * @param ClearTextPassword|string $newPassword
+     * @param \Stringable|string $newPassword
      *
      * @throws Exception\AuthException
      * @throws Exception\FirebaseException
@@ -324,7 +323,7 @@ interface Auth
      * @see https://firebase.google.com/docs/reference/rest/auth#section-confirm-reset-password
      *
      * @param string $oobCode the email action code sent to the user's email for resetting the password
-     * @param ClearTextPassword|string $newPassword
+     * @param \Stringable|string $newPassword
      * @param bool $invalidatePreviousSessions Invalidate sessions initialized with the previous credentials
      *
      * @throws ExpiredOobCode
@@ -342,7 +341,7 @@ interface Auth
      * @see https://firebase.google.com/docs/reference/rest/auth#section-confirm-reset-password
      *
      * @param string $oobCode the email action code sent to the user's email for resetting the password
-     * @param ClearTextPassword|string $newPassword
+     * @param \Stringable|string $newPassword
      * @param bool $invalidatePreviousSessions Invalidate sessions initialized with the previous credentials
      *
      * @throws ExpiredOobCode
@@ -398,7 +397,7 @@ interface Auth
 
     /**
      * @param \Stringable|string $email
-     * @param string|ClearTextPassword $clearTextPassword
+     * @param \Stringable|string $clearTextPassword
      *
      * @throws FailedToSignIn
      */